# fern/definition/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Airweave",
    "version": "0.1.0",
    "description": "\n\n## API Groups\nThis API spec only includes the following API groups:\n- **Sources**: API endpoints for discovering available data source connectors and their configuration requirements\n- **Collections**: API endpoints for managing collections - logical groups of data sources that provide unified search capabilities\n- **Source Connections**: API endpoints for managing live connections to data sources. Source connections are the actual configured instances that Airweave uses to sync data from your apps and databases, transforming it into searchable, structured information within collections\n"
  },
  "paths": {
...
    "/collections": {
      "get": {
        "tags": [
          "collections"
        ],
        "summary": "List",
        "description": "List all collections that belong to your organization.",
        "operationId": "list_collections_get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Number of collections to skip for pagination",
              "default": 0,
              "title": "Skip"
            },
            "description": "Number of collections to skip for pagination"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "Maximum number of collections to return (1-1000)",
              "default": 100,
              "title": "Limit"
            },
            "description": "Maximum number of collections to return (1-1000)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Collection"
                  },
                  "title": "Response List Collections Get"
                },
                "examples": {
                  "collections_list": {
                    "summary": "Finance data collection",
                    "value": [
                      {
                        "id": "550e8400-e29b-41d4-a716-446655440000",
                        "name": "Finance Data",
                        "readable_id": "finance-data-ab123",
                        "created_at": "2024-01-15T09:30:00Z",
                        "modified_at": "2024-01-15T14:22:15Z",
                        "organization_id": "org12345-6789-abcd-ef01-234567890abc",
                        "created_by_email": "admin@company.com",
                        "modified_by_email": "finance@company.com",
                        "status": "ACTIVE"
                      }
                    ]
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "collections"
        ],
        "summary": "Create",
        "description": "Create a new collection.\n\nThe newly created collection is initially empty and does not contain any data\nuntil you explicitly add source connections to it.",
        "operationId": "create_collections_post",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Collection"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/collections/{readable_id}": {
      "get": {
        "tags": [
          "collections"
        ],
        "summary": "Get",
        "description": "Retrieve a specific collection by its readable ID.",
        "operationId": "get_collections__readable_id__get",
        "parameters": [
          {
            "name": "readable_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The unique readable identifier of the collection (e.g., 'finance-data-ab123')",
              "title": "Readable Id"
            },
            "description": "The unique readable identifier of the collection (e.g., 'finance-data-ab123')"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Collection"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "collections"
        ],
        "summary": "Delete",
        "description": "Delete a collection and all associated data.\n\nPermanently removes a collection from your organization including all synced data\nfrom the destination systems. All source connections within this collection\nwill also be deleted as part of the cleanup process. This action cannot be undone.",
        "operationId": "delete_collections__readable_id__delete",
        "parameters": [
          {
            "name": "readable_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The unique readable identifier of the collection to delete",
              "title": "Readable Id"
            },
            "description": "The unique readable identifier of the collection to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Collection"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
    "/collections/{readable_id}/refresh_all": {
      "post": {
        "tags": [
          "collections"
        ],
        "summary": "Refresh All Source Connections",
        "description": "Trigger data synchronization for all source connections in the collection.\n\nThe sync jobs run asynchronously in the background, so this endpoint\nreturns immediately with job details that you can use to track progress. You can\nmonitor the status of individual data synchronization using the source connection\nendpoints.",
        "operationId": "refresh_all_source_connections_collections__readable_id__refresh_all_post",
        "parameters": [
          {
            "name": "readable_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The unique readable identifier of the collection to refresh",
              "title": "Readable Id"
            },
            "description": "The unique readable identifier of the collection to refresh"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SourceConnectionJob"
                  },
                  "title": "Response Refresh All Source Connections Collections  Readable Id  Refresh All Post"
                },
                "examples": {
                  "sync_job_history": {
                    "summary": "Multiple sync jobs triggered",
                    "value": [
                      {
                        "source_connection_id": "550e8400-e29b-41d4-a716-446655440000",
                        "id": "987fcdeb-51a2-43d7-8f3e-1234567890ab",
                        "organization_id": "org12345-6789-abcd-ef01-234567890abc",
                        "created_by_email": "engineering@company.com",
                        "modified_by_email": "engineering@company.com",
                        "created_at": "2024-01-15T14:00:00Z",
                        "modified_at": "2024-01-15T14:05:22Z",
                        "status": "completed",
                        "entities_inserted": 45,
                        "entities_updated": 12,
                        "entities_deleted": 3,
                        "entities_kept": 234,
                        "entities_skipped": 8,
                        "entities_encountered": {
                          "issues": 67,
                          "pull_requests": 23,
                          "commits": 156,
                          "releases": 12,
                          "readme_files": 8
                        },
                        "started_at": "2024-01-15T14:00:15Z",
                        "completed_at": "2024-01-15T14:05:22Z"
                      }
                    ]
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Collection": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Human-readable display name for the collection."
          },
          "readable_id": {
            "type": "string",
            "title": "Readable Id",
            "description": "URL-safe unique identifier used in API endpoints. This becomes non-optional once the collection is created."
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Unique system identifier for the collection. This UUID is generated automatically and used for internal references."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Timestamp when the collection was created (ISO 8601 format)."
          },
          "modified_at": {
            "type": "string",
            "format": "date-time",
            "title": "Modified At",
            "description": "Timestamp when the collection was last modified (ISO 8601 format)."
          },
          "organization_id": {
            "type": "string",
            "format": "uuid",
            "title": "Organization Id",
            "description": "Identifier of the organization that owns this collection. Collections are isolated per organization."
          },
          "created_by_email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By Email",
            "description": "Email address of the user who created this collection."
          },
          "modified_by_email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modified By Email",
            "description": "Email address of the user who last modified this collection."
          },
          "status": {
            "$ref": "#/components/schemas/CollectionStatus",
            "description": "Current operational status of the collection:<br/>\u2022 **NEEDS_SOURCE**: Collection has no authenticated connections, or connections exist but haven't synced yet<br/>\u2022 **ACTIVE**: At least one connection has completed a sync or is currently syncing<br/>\u2022 **ERROR**: All connections have failed their last sync",
            "default": "NEEDS SOURCE"
          }
        },
        "type": "object",
        "required": [
          "name",
          "readable_id",
          "id",
          "created_at",
          "modified_at",
          "organization_id"
        ],
        "title": "Collection",
        "description": "Complete collection representation returned by the API.\n\nThis schema includes all collection metadata plus computed status information\nbased on the health and state of associated source connections.",
        "examples": [
          {
            "created_at": "2024-01-15T09:30:00Z",
            "created_by_email": "admin@company.com",
            "id": "550e8400-e29b-41d4-a716-446655440000",
            "modified_at": "2024-01-15T14:22:15Z",
            "modified_by_email": "finance@company.com",
            "name": "Finance Data",
            "organization_id": "org12345-6789-abcd-ef01-234567890abc",
            "readable_id": "finance-data-ab123",
            "status": "ACTIVE"
          }
        ]
      },
      "CollectionCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 64,
            "minLength": 4,
            "title": "Name",
            "description": "Human-readable display name for the collection. This appears in the UI and should clearly describe the data contained within (e.g., 'Finance Data').",
            "examples": [
              "Finance Data",
              "Customer Support",
              "Marketing Analytics"
            ]
          },
          "readable_id": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Readable Id",
            "description": "URL-safe unique identifier used in API endpoints. Must contain only lowercase letters, numbers, and hyphens. If not provided, it will be automatically generated from the collection name with a random suffix for uniqueness (e.g., 'finance-data-ab123').",
            "examples": [
              "finance-data-ab123",
              "customer-support-xy789",
              "marketing-analytics-cd456"
            ]
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CollectionCreate",
        "description": "Schema for creating a new collection.\n\nCollections serve as logical containers for organizing related data sources.\nOnce created, you can add source connections to populate the collection with data\nfrom various sources like databases, APIs, and file systems.",
        "examples": [
          {
            "name": "Finance Data",
            "readable_id": "finance-data-reports"
          },
          {
            "name": "Customer Support"
          }
        ]
      },
      "CollectionStatus": {
        "type": "string",
        "enum": [
          "ACTIVE",
          "NEEDS SOURCE",
          "ERROR"
        ],
        "title": "CollectionStatus",
        "description": "Collection status enum."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SourceConnectionJob": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "source_connection_id": {
            "type": "string",
            "format": "uuid",
            "title": "Source Connection Id"
          },
          "status": {
            "$ref": "#/components/schemas/SyncJobStatus"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "entities_inserted": {
            "type": "integer",
            "title": "Entities Inserted",
            "default": 0
          },
          "entities_updated": {
            "type": "integer",
            "title": "Entities Updated",
            "default": 0
          },
          "entities_deleted": {
            "type": "integer",
            "title": "Entities Deleted",
            "default": 0
          },
          "entities_failed": {
            "type": "integer",
            "title": "Entities Failed",
            "default": 0
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "error_details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Details"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source_connection_id",
          "status"
        ],
        "title": "SourceConnectionJob",
        "description": "Individual sync job for a source connection."
      },
...
      "SyncJobStatus": {
        "type": "string",
        "enum": [
          "created",
          "pending",
          "running",
          "completed",
          "failed",
          "cancelling",
          "cancelled"
        ],
        "title": "SyncJobStatus",
        "description": "Sync job status enum."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key for authentication"
      }
    }
  },
...
  "servers": [
    {
      "url": "https://api.airweave.ai",
      "description": "Production",
      "x-fern-server-name": "Production"
    },
    {
      "url": "http://localhost:8001",
      "description": "Local",
      "x-fern-server-name": "Local"
    }
  ]
}